Simplified approach to just detect when gameState is not being received
