+++ game_frontend/src/redux/features/Game/epics.js
@@ -1,8 +1,15 @@
 import actions from './actions'
 import types from './types'
 import { of } from 'rxjs'
-import { map, mergeMap, catchError, switchMap, first, mapTo } from 'rxjs/operators'
+import { map, mergeMap, catchError, switchMap, first, mapTo, debounceTime } from 'rxjs/operators'
 import { ofType } from 'redux-observable'
+
+const timeoutEpic = (action$) =>
+  action$.pipe(
+    ofType(types.SOCKET_GAME_STATE_RECEIVED),
+    debounceTime(12000),
+    map(action => actions.setTimeout())
+  )
 
 const getConnectionParametersEpic = (action$, state$, { api: { get } }) => action$.pipe(
   ofType(types.SOCKET_CONNECT_TO_GAME_REQUEST),
@@ -62,5 +69,6 @@ export default {
   connectToGameEpic,
   gameLoadedEpic,
   sendGameStateEpic,
-  sendAvatarIDEpic
+  sendAvatarIDEpic,
+  timeoutEpic
 }
