+++ game_frontend/src/redux/features/Game/reducers.js
@@ -2,6 +2,11 @@ import types from './types'
 
 const gameReducer = (state = {}, action) => {
   switch (action.type) {
+    case types.SET_TIMEOUT:
+      return {
+        ...state,
+        timeoutStatus: true
+      }
     case types.SOCKET_GAME_STATE_RECEIVED:
       return {
         ...state,
