+++ game_frontend/src/redux/features/Game/types.js
@@ -17,6 +17,8 @@ const SNACKBAR_FOR_AVATAR_FEEDBACK_SHOWN = 'features/Game/SNACKBAR_FOR_AVATAR_FE
 
 const GAME_DATA_LOADED = 'features/Game/GAME_DATA_LOADED'
 
+const SET_TIMEOUT = 'features/Editor/TIMEOUT'
+
 export default {
   SOCKET_CONNECT_TO_GAME_REQUEST,
   SOCKET_CONNECT_TO_GAME_FAIL,
@@ -29,5 +31,6 @@ export default {
   UNITY_SEND_AVATAR_ID_FAIL,
   SOCKET_FEEDBACK_AVATAR_UPDATED,
   SNACKBAR_FOR_AVATAR_FEEDBACK_SHOWN,
-  GAME_DATA_LOADED
+  GAME_DATA_LOADED,
+  SET_TIMEOUT
 }
